requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
WTForms==3.1.1
validators==0.22.0
python-dotenv==1.0.0
gunicorn==21.2.0
psutil==5.9.6